Addresses the feedbacks. Refactoring the common code blocks into demo_util.js file.
