+++ posenet/demos/demo_util.js
@@ -17,24 +17,44 @@
 import * as posenet from '@tensorflow-models/posenet';
 import * as tf from '@tensorflow/tfjs';
 
-export const TRY_RESNET50_BUTTON_TEXT = 'Try ResNet50'
-
 const color = 'aqua';
 const boundingBoxColor = 'red';
 const lineWidth = 2;
 
-export function setDatGuiPropertyCss(
-    propertyName, liCssString, spanCssString = '') {
+export const tryResNetButtonName = 'tryResNetButton';
+export const tryResNetButtonText = '[New] Try ResNet50';
+const tryResNetButtonTextCss = 'width:100%;text-decoration:underline;';
+const tryResNetButtonBackgroundCss = 'background:#e61d5f;';
+
+function isAndroid() {
+  return /Android/i.test(navigator.userAgent);
+}
+
+function isiOS() {
+  return /iPhone|iPad|iPod/i.test(navigator.userAgent);
+}
+
+export function isMobile() {
+  return isAndroid() || isiOS();
+}
+
+function setDatGuiPropertyCss(propertyText, liCssString, spanCssString = '') {
   var spans = document.getElementsByClassName('property-name');
   for (var i = 0; i < spans.length; i++) {
     var text = spans[i].textContent || spans[i].innerText;
-    if (text == propertyName) {
+    if (text == propertyText) {
       spans[i].parentNode.parentNode.style = liCssString;
       if (spanCssString !== '') {
         spans[i].style = spanCssString;
       }
     }
   }
+}
+
+export function updateTryResNetButtonDatGuiCss() {
+  setDatGuiPropertyCss(
+      tryResNetButtonText, tryResNetButtonBackgroundCss,
+      tryResNetButtonTextCss);
 }
 
 function toTuple({y, x}) {
